Reorganized file structure for PyGame RC example and added requirements.txt and README
