+++ examples/pygame-rc/README.md
@@ -1,0 +1,9 @@
+# PyGame Tello RC
+
+This script runs PyGame and shows the camera feed coming from a connected Tello Drone. Users can also send RC commands to the Tello Drone to control it in real time.
+
+The controls are as follows:
+- A/D - Move Left/Right (Roll)
+- W/S - Move Forward/Backward (Pitch)
+- Q/E - Turn Left/Right (Yaw)
+- R/F - Fly Up/Down
